시큐리티 적용

+++ src/main/java/com/example/enjoytrip/account/service/AccountServiceImpl.java
@@ -8,6 +8,9 @@ import com.example.enjoytrip.exception.CustomException;
 import com.example.enjoytrip.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,10 +21,21 @@ public class AccountServiceImpl implements AccountService{
 
     private final AccountDao accountDao;
     private final ModelMapper modelMapper;
+    private final PasswordEncoder passwordEncoder;
 
     @Transactional
     @Override
     public int delete(Integer accountId) {
+        AccountResponseDto account = findById(accountId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!authentication.isAuthenticated()){
+            throw new CustomException(ErrorCode.Unauthorized);
+        }
+        if(!authentication.getPrincipal().equals(account.getAccountEmail())){
+            throw new CustomException(ErrorCode.Forbidden);
+        }
+
+
         return accountDao.delete(accountId);
     }
 
@@ -32,6 +46,7 @@ public class AccountServiceImpl implements AccountService{
         if(byEmail != null){
             throw new CustomException(ErrorCode.DuplicateUserEmail);
         }
+        accountRequestDto.setAccountPassword(passwordEncoder.encode(accountRequestDto.getAccountPassword()));
         Account account = modelMapper.map(accountRequestDto, Account.class);
         return accountDao.join(account);
     }
@@ -46,6 +61,14 @@ public class AccountServiceImpl implements AccountService{
     @Transactional
     @Override
     public int update(Account account) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!authentication.isAuthenticated()){
+            throw new CustomException(ErrorCode.Unauthorized);
+        }
+        if(!authentication.getPrincipal().equals(account.getAccountEmail())){
+            throw new CustomException(ErrorCode.Forbidden);
+        }
+
         return accountDao.update(account);
     }
 }
